pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dendroviz"
version = "0.1.0"
description = "Generate dendrogram scaffolding CSV and SVG from rooted tree data."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
keywords = ["dendrogram", "tree", "svg", "csv", "visualization"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: Scientific/Engineering :: Visualization",
]
dependencies = []

[project.optional-dependencies]
newick = [
  "biopython>=1.83",
]
dev = [
  "build>=1.2.2",
  "mypy>=1.10",
  "ruff>=0.6.9",
]

[project.scripts]
dendroviz = "dendroviz.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "B"]

[tool.mypy]
python_version = "3.10"
strict = true
packages = ["dendroviz"]
